API: Move all code generation into cpy package.
All the code generation related code is now in the `pysph.cpy` package.
The code generation features will be further worked on and this package
may become a separate package later on.  Changed all other imports so
tests pass.
