+++ pysph/cpy/capture_stream.py
