+++ pysph/cpy/cython_generator.py
@@ -18,9 +18,8 @@ import types
 from mako.template import Template
 import numpy
 
-from pysph.base.ast_utils import get_assigned, has_return
 from pysph.base.config import get_config
-
+from .ast_utils import get_assigned, has_return
 
 logger = logging.getLogger(__name__)
