+++ pysph/cpy/tests/test_cython_generator.py
@@ -6,8 +6,8 @@ from math import pi, sin
 import numpy as np
 
 from pysph.base.config import get_config, set_config
-from pysph.base.cython_generator import (CythonGenerator, CythonClassHelper,
-                                         KnownType, all_numeric, declare)
+from ..cython_generator import (CythonGenerator, CythonClassHelper,
+                                KnownType, all_numeric, declare)
 
 
 class BasicEq:
