+++ pysph/sph/acceleration_eval_opencl_helper.py
@@ -86,11 +86,11 @@ import pyopencl.tools  # noqa: 401
 
 from pysph.base.config import get_config
 from pysph.base.utils import is_overloaded_method
-from pysph.base.ext_module import get_platform_dir
+from pysph.cpy.ext_module import get_platform_dir
 from pysph.base.opencl import (profile_kernel, get_context, get_queue,
                                DeviceHelper)
-from pysph.base.translator import (CStructHelper, OpenCLConverter,
-                                   ocl_detect_type)
+from pysph.cpy.translator import (CStructHelper, OpenCLConverter,
+                                  ocl_detect_type)
 
 from .equation import get_predefined_types, KnownType
 from .acceleration_eval_cython_helper import (
