+++ pysph/sph/gas_dynamics/basic.py
@@ -1,6 +1,6 @@
 """Basic equations for Gas-dynamics"""
 
-from pysph.base.cython_generator import declare
+from pysph.cpy.api import declare
 from pysph.base.reduce_array import serial_reduce_array, parallel_reduce_array
 from pysph.sph.equation import Equation
 from math import sqrt, exp, log
