+++ pysph/sph/integrator_opencl_helper.py
@@ -9,7 +9,7 @@ import numpy as np
 
 from pysph.base.config import get_config
 from pysph.base.opencl import profile_kernel
-from pysph.base.translator import OpenCLConverter
+from pysph.cpy.translator import OpenCLConverter
 from .equation import get_array_names
 from .integrator_cython_helper import IntegratorCythonHelper
 from .acceleration_eval_opencl_helper import (get_kernel_definition,
